+++ template_test.go
@@ -1,16 +1,19 @@
 package textwire
 
 import (
+	"net/http/httptest"
+	"strings"
 	"testing"
 
 	"github.com/textwire/textwire/v3/config"
 	"github.com/textwire/textwire/v3/fail"
+	"github.com/textwire/textwire/v3/file"
+	"github.com/textwire/textwire/v3/object"
 )
 
 func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
-	path, err := getFullPath("")
-	path += "/textwire/testdata/bad/"
-
+	absPath, err := file.ToFullPath("")
+	absPath += "/textwire/testdata/bad/"
 	if err != nil {
 		t.Errorf("unexpected error: %s", err)
 		return
@@ -25,7 +28,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "use-inside-tpl",
 			err: fail.New(
 				1,
-				path+"use-inside-tpl/index.tw",
+				absPath+"use-inside-tpl/index.tw",
 				"evaluator",
 				fail.ErrUseStmtNotAllowed,
 			),
@@ -35,7 +38,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "unknown-named-slot",
 			err: fail.New(
 				2,
-				path+"unknown-named-slot/index.tw",
+				absPath+"unknown-named-slot/index.tw",
 				"parser",
 				fail.ErrSlotNotDefined,
 				"unknown",
@@ -47,7 +50,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "unknown-default-slot",
 			err: fail.New(
 				2,
-				path+"unknown-default-slot/index.tw",
+				absPath+"unknown-default-slot/index.tw",
 				"parser",
 				fail.ErrDefaultSlotNotDefined,
 				"book",
@@ -58,7 +61,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "duplicate-slot",
 			err: fail.New(
 				2,
-				path+"duplicate-slot/index.tw",
+				absPath+"duplicate-slot/index.tw",
 				"parser",
 				fail.ErrDuplicateSlot,
 				"content",
@@ -71,7 +74,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "duplicate-default-slot",
 			err: fail.New(
 				2,
-				path+"duplicate-default-slot/index.tw",
+				absPath+"duplicate-default-slot/index.tw",
 				"parser",
 				fail.ErrDuplicateDefaultSlot,
 				2,
@@ -83,7 +86,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "unknown-comp",
 			err: fail.New(
 				9,
-				path+"unknown-comp/index.tw",
+				absPath+"unknown-comp/index.tw",
 				"template",
 				fail.ErrUndefinedComponent,
 				"unknown-name",
@@ -94,7 +97,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "undefined-insert",
 			err: fail.New(
 				5,
-				path+"undefined-insert/index.tw",
+				absPath+"undefined-insert/index.tw",
 				"parser",
 				fail.ErrAddMatchingReserve,
 				"some-name",
@@ -106,7 +109,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "duplicate-inserts",
 			err: fail.New(
 				4,
-				path+"duplicate-inserts/index.tw",
+				absPath+"duplicate-inserts/index.tw",
 				"parser",
 				fail.ErrDuplicateInserts,
 				"title",
@@ -117,7 +120,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "undefined-var-in-comp",
 			err: fail.New(
 				1,
-				path+"undefined-var-in-comp/hero.tw",
+				absPath+"undefined-var-in-comp/hero.tw",
 				"evaluator",
 				fail.ErrVariableIsUndefined,
 				"undefinedVar",
@@ -128,7 +131,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "undefined-var-in-use",
 			err: fail.New(
 				8,
-				path+"undefined-var-in-use/base.tw",
+				absPath+"undefined-var-in-use/base.tw",
 				"evaluator",
 				fail.ErrVariableIsUndefined,
 				"undefinedVar",
@@ -139,7 +142,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "undefined-use",
 			err: fail.New(
 				1,
-				path+"undefined-use/index.tw",
+				absPath+"undefined-use/index.tw",
 				"parser",
 				fail.ErrUseStmtMissingLayout,
 				"undefined-layout",
@@ -150,7 +153,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "undefined-var-in-nested-comp",
 			err: fail.New(
 				1,
-				path+"undefined-var-in-nested-comp/second.tw",
+				absPath+"undefined-var-in-nested-comp/second.tw",
 				"evaluator",
 				fail.ErrVariableIsUndefined,
 				"name",
@@ -161,7 +164,7 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 			dir: "var-in-layout",
 			err: fail.New(
 				1,
-				path+"var-in-layout/layout.tw",
+				absPath+"var-in-layout/layout.tw",
 				"evaluator",
 				fail.ErrVariableIsUndefined,
 				"fullName",
@@ -170,14 +173,14 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 		},
 		{
 			dir:  "duplicate-use",
-			err:  fail.New(2, path+"duplicate-use/index.tw", "parser", fail.ErrOnlyOneUseDir),
+			err:  fail.New(2, absPath+"duplicate-use/index.tw", "parser", fail.ErrOnlyOneUseDir),
 			data: nil,
 		},
 		{
 			dir: "inserts-without-use",
 			err: fail.New(
 				4,
-				path+"inserts-without-use/index.tw",
+				absPath+"inserts-without-use/index.tw",
 				"evaluator",
 				fail.ErrInsertRequiresUse,
 				"title",
@@ -221,66 +224,159 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 
 func TestNewTemplate(t *testing.T) {
 	cases := []struct {
-		dirName  string
-		viewName string
-		data     map[string]any
+		conf *config.Config
+		view string
+		data map[string]any
+		dir  string
 	}{
-		{"no-stmts", "index", nil},
-		{"with-inserts", "index", nil},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "no-stmts"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "with-inserts"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "use-inside-if"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "with-comp"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "with-inserts-and-html"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "with-comp-no-args"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "insert-is-optional"},
+		{conf: &config.Config{}, view: "index", data: nil, dir: "use-with-comp-inside"},
+		{conf: &config.Config{}, view: "home", data: nil, dir: "comp-in-other-comp"},
 		{
-			"without-use",
-			"index",
-			map[string]any{
+			conf: &config.Config{},
+			view: "index",
+			data: map[string]any{
 				"pageTitle": "Test Page",
 				"NAME_1":    "Anna Korotchaeva",
 				"name_2":    "Serhii Cho",
 			},
+			dir: "without-use",
 		},
-		{"loops", "index", map[string]any{"names": []string{"Anna", "Serhii", "Vladimir"}}},
 		{
-			"with-each-and-comp",
-			"views/index",
-			map[string]any{"names": []string{"Anna", "Serhii", "Vladimir"}},
+			conf: &config.Config{},
+			view: "index",
+			data: map[string]any{"names": []string{"Anna", "Serhii", "Vladimir"}},
+			dir:  "loops",
 		},
-		{"use-inside-if", "index", nil},
-		{"with-comp", "index", nil},
-		{"with-inserts-and-html", "index", nil},
 		{
-			"with-comp-and-slots",
-			"index",
-			map[string]any{"name": "Anna", "age": 20},
+			conf: &config.Config{},
+			view: "views/index",
+			data: map[string]any{"names": []string{"Anna", "Serhii", "Vladimir"}},
+			dir:  "with-each-and-comp",
 		},
-		{"with-comp-no-args", "index", nil},
-		{"insert-is-optional", "index", nil},
-		{"use-with-comp-inside", "index", nil},
-		{"comp-in-other-comp", "home", nil},
+		{
+			conf: &config.Config{},
+			view: "index",
+			data: map[string]any{"name": "Anna", "age": 20},
+			dir:  "with-comp-and-slots",
+		},
 	}
 
 	for _, tc := range cases {
-		t.Run(tc.dirName, func(t *testing.T) {
-			tpl, err := NewTemplate(&config.Config{
-				TemplateDir: "textwire/testdata/good/before/" + tc.dirName,
-			})
+		t.Run(tc.dir, func(t *testing.T) {
+			tc.conf.TemplateDir = "textwire/testdata/good/before/" + tc.dir
+			tpl, err := NewTemplate(tc.conf)
+			if err != nil {
+				t.Fatalf("Error creating template: %q", err)
+			}
 
+			actual, failure := tpl.String(tc.view, tc.data)
+			if failure != nil {
+				t.Fatalf("Error evaluating template: %q", failure)
+			}
+
+			expect, err := readFile("textwire/testdata/good/expected/" + tc.dir + ".html")
+			if err != nil {
+				t.Fatalf("Error reading file. Error: %s", err)
+			}
+
+			if actual != expect {
+				t.Fatalf("Wrong result. Expect:\n'%s'\ngot:\n'%s'", expect, actual)
+			}
+		})
+	}
+}
+
+func TestTemplateResponse(t *testing.T) {
+	absPath, err := file.ToFullPath("")
+	absPath += "/textwire/testdata/good/before/"
+	if err != nil {
+		t.Errorf("unexpected error: %s", err)
+		return
+	}
+
+	cases := []struct {
+		name string
+		conf *config.Config
+		view string
+		data map[string]any
+		dir  string
+		err  *fail.Error
+	}{
+		{
+			name: "Should show custom error page",
+			conf: &config.Config{ErrorPagePath: "custom-error-page"},
+			view: "home",
+			dir:  "prod-error-page",
+			data: map[string]any{"arr": "some string"},
+			err: fail.New(
+				2,
+				absPath+"prod-error-page/home.tw",
+				"parser",
+				fail.ErrEachDirWithNonArrArg,
+				object.STR_OBJ,
+			),
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			tc.conf.TemplateDir += "textwire/testdata/good/before/" + tc.dir
+			tpl, err := NewTemplate(tc.conf)
 			if err != nil {
 				t.Errorf("Error creating template: %q", err)
 				return
 			}
 
-			actual, failure := tpl.String(tc.viewName, tc.data)
-			if failure != nil {
-				t.Fatalf("Error evaluating template: %q", failure)
-				return
-			}
-
-			expect, err := readFile("textwire/testdata/good/expected/" + tc.dirName + ".html")
+			expect, err := readFile("textwire/testdata/good/expected/" + tc.dir + ".html")
 			if err != nil {
 				t.Fatalf("Error reading file. Error: %s", err)
-				return
+			}
+
+			rr := httptest.NewRecorder()
+			respErr := tpl.Response(rr, tc.view, tc.data)
+			actual := rr.Body.String()
+
+			if tc.err != nil {
+				if respErr == nil {
+					t.Fatalf("We expect error from response, got nil")
+				}
+
+				if tc.err.String() != respErr.Error() {
+					t.Fatalf("Wrong error message! Expect:\n%q\ngot:\n%q", tc.err, respErr)
+				}
 			}
 
 			if actual != expect {
 				t.Fatalf("Wrong result. Expect:\n'%s'\ngot:\n'%s'", expect, actual)
+			}
+
+			// Make sure you don't see error in actual response without debug mode
+			if !tc.conf.DebugMode && respErr != nil {
+				if contains := strings.Contains(actual, respErr.Error()); contains {
+					t.Fatalf(
+						"Actual string should not contain error message. Actual:\n'%s'\nError msg:\n'%s'",
+						actual,
+						respErr,
+					)
+				}
+			}
+
+			// Make sure you see error in actual response with debug mode enabled
+			if tc.conf.DebugMode && respErr != nil {
+				if contains := strings.Contains(actual, respErr.Error()); !contains {
+					t.Fatalf(
+						"Actual string should contain error message. Actual:\n'%s'\nError msg:\n'%s'",
+						actual,
+						respErr,
+					)
+				}
 			}
 		})
 	}